third_party/blink/web_tests/fast/frames/srcdoc/srcdoc-cannot-be-in-qurks-mode-expected.txt

ALERT: CSS1Compat
Normally srcdoc documents default to standards mode, but they can end up in quirks mode with a sufficiently nutty DocType. 
